+++ packages/angular/src/ng-add-setup-project/files/src/app/init.service.ts
@@ -1,16 +1,23 @@
+import {Router} from '@angular/router';
 import {Injectable, APP_INITIALIZER} from '@angular/core';
-import {environment} from 'src/environments/environment';
+import {environment} from '../environments/environment';
 import {EngineService} from './engine.service';
 
 @Injectable({
   providedIn: 'root',
 })
 export class InitService {
-  constructor(private engineService: EngineService) {}
+  constructor(private engineService: EngineService, private route: Router) {}
 
   async init() {
     const res = await fetch(environment.tokenEndpoint);
-    const {token} = await res.json();
+    const data = await res.json();
+    const token = data.token;
+
+    if (!token) {
+      this.route.navigate(['error'], {state: {errorMessage: data.message}});
+      return;
+    }
     this.engineService.init(token);
   }
 }