+++ packages/angular/src/ng-add-setup-project/files/src/environments/environment.d.ts
@@ -1,0 +1,6 @@
+export declare const environment: {
+  production: boolean;
+  tokenEndpoint: string;
+  platformUrl: string;
+  organizationId: string;
+};