+++ packages/cli/src/lib/decorators/preconditions/testsUtils/utils.ts
@@ -1,0 +1,10 @@
+import {Command} from '@oclif/command';
+
+export const getFakeCommand = (): Command => {
+  const fakeCommand = {
+    id: 'foo',
+    warn: jest.fn(),
+  };
+
+  return (fakeCommand as unknown) as Command;
+};